Validation by users added

+++ src/main/java/com/thing/web/servlet/RemoveProductServlet.java
@@ -1,6 +1,7 @@
 package com.thing.web.servlet;
 
 import com.thing.service.ProductService;
+import com.thing.service.impl.SecurityService;
 
 import javax.servlet.ServletException;
 import javax.servlet.http.HttpServlet;
@@ -10,16 +11,30 @@ import java.io.IOException;
 
 public class RemoveProductServlet extends HttpServlet {
     private ProductService productService;
+    private SecurityService securityService;
 
     @Override
     protected void doPost(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
-        int id = Integer.parseInt(req.getParameter("id"));
-        productService.removeProduct(id);
+        if (securityService.containsCoockie(req.getCookies())){
+            if (securityService.isAdmin()) {
+                int id = Integer.parseInt(req.getParameter("id"));
+                productService.removeProduct(id);
 
-        resp.sendRedirect("/products");
+                resp.sendRedirect("/products");
+            } else {
+                resp.sendRedirect("/guest");
+            }
+        } else {
+            resp.sendRedirect("/login");
+        }
+
     }
 
     public void setProductService(ProductService productService) {
         this.productService = productService;
+    }
+
+    public void setSecurityService(SecurityService securityService) {
+        this.securityService = securityService;
     }
 }
